Almost have the wireless module working...

+++ FYP_Hexabot_ATSAM_S70/FYP_Hexabot_ATSAM_S70/src/Hexabot/Hexabot_Cmd.h
@@ -36,5 +36,5 @@ void cmdRelaxAll();
 
 void cmdRXen();
 
-int cmdDWMreadRX(char* buffer) ;
+uint64_t cmdDWMreadRX(char* buffer) ;
 #endif /* HEXABOT_CMD_H_ */
